[Update] 에너미 시야각 & 스테이터스 저장 기능 추가

+++ Assets/3.Script/UI/Character/Status.cs
@@ -9,5 +9,5 @@ public class Status : ScriptableObject
     public float curMp;
     public float damage;
     public float def;
-    public int exp;
+    public float exp;
 }
